app: thread HeapAllocator through core module Init calls
Aos core library APIs (crypto provider, PKCS11 manager, cert loader,
cert handler, space allocators, image manager, launcher, monitoring,
network manager, node config handler, etc.) now take an allocator as
part of their redesigned Init signatures, so every app entry point
(cm, iam, mp, sm) gains a HeapAllocator member and passes it through.

Also refactors the IAM app to move its Init/Start/Stop logic out of
App into a new AosCore class, matching the pattern already used by
cm and sm, which keeps App as a thin Poco::Util::ServerApplication
wrapper around AosCore.

Updates the affected unit tests to construct a HeapAllocator and
pass it to the library calls whose signatures changed, and fixes
FSPlatform::GetDirSize to call the renamed common::utils::CalculateSize
helper instead of the removed aos::fs::CalculateSize.
