+++ src/iam/app/CMakeLists.txt
@@ -10,7 +10,7 @@ set(TARGET_NAME app)
 # Sources
 # ######################################################################################################################
 
-set(SOURCES app.cpp main.cpp)
+set(SOURCES aoscore.cpp app.cpp main.cpp)
 
 # ######################################################################################################################
 # Libraries
